Implementa criacao e busca de instituicoes

+++ src/main/java/com/recrutaibackend/dto/InstitutionRequest.java
@@ -25,7 +25,7 @@ public record InstitutionRequest(
         String industry,
 
         @NotNull
-        Integer companySizeId,
+        Short companySizeId,
 
         @Valid
         AddressRequest headquarters,
